+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Upgraded dependencies
 - Larger map with interaction
 
+### Added:
+- Display size of downloaded whisper models
+
 ## [0.8.356] - 2023-05-21
 ### Changed:
 - Improve header style on settings page using Slivers